add application logic
the cpu logic implemetned is the following: cpu0 receives and timer
interrupt and sends an ipi to the next cpu. When a cpu receives an ipi
it send the same ipi to next cpu. Also, all cpus enable the uart rx
interrupt which should be triggered by sending a single character

this includes makes part of the irq and timer interface arch an plat
independent